+++ vite.config.ts
@@ -16,6 +16,7 @@ export default defineConfig({
                 userguide: resolve(__dirname, 'src/pages/userguide/index.html'),
                 privacy: resolve(__dirname, 'src/pages/privacy/index.html'),
                 updates: resolve(__dirname, 'src/pages/updates/index.html'),
+                accuracy: resolve(__dirname, 'src/pages/accuracy/index.html'),
                 background: resolve(__dirname, 'src/background/index.ts'),
                 content: resolve(__dirname, 'src/content/index.ts'),
             },